requirements.txt
torch>=1.13.0
torchvision>=0.14.0
timm>=0.6.0
Pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.5.0
requests>=2.25.0
matplotlib>=3.5.0
huggingface-hub>=0.15.0
gdown>=4.0.0